feat: Restructuring Control panel files and adding feedback dialogs
