+++ package.json
@@ -27,6 +27,7 @@
     "react-router-dom": "^6.25.1",
     "react-scripts": "5.0.1",
     "styled-components": "^6.1.12",
+    "sweetalert2": "^11.12.3",
     "web-vitals": "^2.1.4",
     "yup": "^1.4.0"
   },
